+++ src/styles.d.ts
@@ -2,3 +2,8 @@ declare module '@advanced-chat/components/styles' {
   const css: string
   export default css
 }
+
+declare module '@advanced-chat/components/web-component/styles' {
+  const css: string
+  export default css
+}