+++ vite.web-component.config.ts
@@ -4,7 +4,7 @@ import path from 'node:path'
 import vue from '@vitejs/plugin-vue'
 import { defineConfig } from 'vite'
 
-export default defineConfig({
+export default defineConfig(({ mode }) => ({
   define: {
     'process.env.NODE_ENV': JSON.stringify('production'),
   },
@@ -22,11 +22,15 @@ export default defineConfig({
     emptyOutDir: false,
     cssCodeSplit: false,
     lib: {
-      entry: path.resolve(__dirname, 'src/web-component.ts'),
-      name: 'VueAdvancedChatElement',
+      entry: path.resolve(
+        __dirname,
+        mode === 'core' ? 'src/web-component-core.ts' : 'src/web-component.ts',
+      ),
+      name: 'AdvancedChatComponentsElement',
       formats: ['es'],
-      fileName: () => 'vue-advanced-chat.js',
-      cssFileName: 'vue-advanced-chat',
+      fileName: () =>
+        mode === 'core' ? 'advanced-chat-components-core.js' : 'advanced-chat-components.js',
+      cssFileName: 'advanced-chat-components',
     },
   },
-})
+}))